Rename as template
