+++ lib/helpers/index.js
@@ -6,7 +6,7 @@ const broadcast = require('./broadcast');
 const request = require('./request');
 const replies = require('./replies');
 const slack = require('./slack');
-const templates = require('./templates');
+const template = require('./template');
 const twilio = require('./twilio');
 
 // TODO: Add "Helper" postfix to prevent name confusion with npm modules like twilio.
@@ -17,6 +17,6 @@ module.exports = {
   request,
   replies,
   slack,
-  templates,
+  template,
   twilio,
 };
