+++ lib/helpers/replies.js
@@ -154,7 +154,7 @@ module.exports.rivescriptReply = function (req, res, messageText) {
  */
 function sendGambitConversationsTemplate(req, res, template) {
   logger.debug('sendGambitConversationsTemplate', { template });
-  const text = helpers.templates.getTextForTemplate(template);
+  const text = helpers.template.getTextForTemplate(template);
   return sendReply(req, res, text, template);
 }
