+++ lib/helpers/template.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const logger = require('../../lib/logger');
-const config = require('../../config/lib/helpers/templates');
+const config = require('../../config/lib/helpers/template');
 
 module.exports = {
   getTextForTemplate: function getTextForTemplate(templateName) {
