+++ lib/middleware/receive-message/campaign-continue.js
@@ -4,7 +4,7 @@ const helpers = require('../../helpers');
 
 module.exports = function askContinueTemplate() {
   return (req, res) => {
-    if (helpers.templates.isGambitCampaignsTemplate(req.lastOutboundTemplate)) {
+    if (helpers.template.isGambitCampaignsTemplate(req.lastOutboundTemplate)) {
       return helpers.replies.continueCampaign(req, res);
     }
