+++ lib/middleware/receive-message/parse-ask-continue-answer.js
@@ -4,7 +4,7 @@ const helpers = require('../../helpers');
 
 module.exports = function parseAskContinueResponse() {
   return (req, res, next) => {
-    const asked = helpers.templates.isAskContinueTemplate(req.lastOutboundTemplate);
+    const asked = helpers.template.isAskContinueTemplate(req.lastOutboundTemplate);
     if (!asked) {
       return next();
     }
