+++ lib/middleware/receive-message/parse-ask-signup-answer.js
@@ -4,7 +4,7 @@ const helpers = require('../../helpers');
 
 module.exports = function parseAskSignupResponse() {
   return (req, res, next) => {
-    const asked = helpers.templates.isAskSignupTemplate(req.lastOutboundTemplate);
+    const asked = helpers.template.isAskSignupTemplate(req.lastOutboundTemplate);
     if (!asked) {
       return next();
     }
