+++ test/lib/lib-helpers/templates.test.js
@@ -12,12 +12,12 @@ chai.should();
 chai.use(sinonChai);
 
 // app modules
-const config = require('../../../config/lib/helpers/templates');
+const config = require('../../../config/lib/helpers/template');
 
 const undefinedTemplateName = 'oathKeeper';
 
 // module to be tested
-const templatesHelper = require('../../../lib/helpers/templates');
+const templateHelper = require('../../../lib/helpers/template');
 
 // sinon sandbox object
 const sandbox = sinon.sandbox.create();
@@ -31,25 +31,25 @@ test.afterEach(() => {
 test('getTextForTemplate should return text for given template', () => {
   const template = 'badWords';
   const templateText = config.templateText[template];
-  const result = templatesHelper.getTextForTemplate(template);
+  const result = templateHelper.getTextForTemplate(template);
   templateText.should.equal(result);
 });
 
 test('getTextForTemplate should return falsy for undefined template', (t) => {
-  t.falsy(templatesHelper.getTextForTemplate(undefinedTemplateName));
+  t.falsy(templateHelper.getTextForTemplate(undefinedTemplateName));
 });
 
 test('isAskContinueTemplate should return boolean', (t) => {
-  t.true(templatesHelper.isAskContinueTemplate('askContinue'));
-  t.falsy(templatesHelper.isAskContinueTemplate(undefinedTemplateName));
+  t.true(templateHelper.isAskContinueTemplate('askContinue'));
+  t.falsy(templateHelper.isAskContinueTemplate(undefinedTemplateName));
 });
 
 test('isAskSignupTemplate should return boolean', (t) => {
-  t.true(templatesHelper.isAskSignupTemplate('askSignup'));
-  t.falsy(templatesHelper.isAskSignupTemplate(undefinedTemplateName));
+  t.true(templateHelper.isAskSignupTemplate('askSignup'));
+  t.falsy(templateHelper.isAskSignupTemplate(undefinedTemplateName));
 });
 
 test('isGambitCampaignsTemplate should return boolean', (t) => {
-  t.true(templatesHelper.isGambitCampaignsTemplate('askQuantity'));
-  t.falsy(templatesHelper.isGambitCampaignsTemplate(undefinedTemplateName));
+  t.true(templateHelper.isGambitCampaignsTemplate('askQuantity'));
+  t.falsy(templateHelper.isGambitCampaignsTemplate(undefinedTemplateName));
 });
